:art: Extract interface
