+++ platforms/platform-googleassistant/src/interfaces.ts
@@ -31,3 +31,14 @@ export interface Context {
 export interface MediaContext {
   progress: string; // Example: "3.5s
 }
+
+export interface GoogleAccountProfile {
+  [key: string]: string | number | boolean;
+
+  email: string;
+  email_verified: boolean;
+  name: string;
+  picture: string;
+  given_name: string;
+  family_name: string;
+}
